pyproject.toml
[tool.poetry]
name = "Basic_website"
version = "0.1.0"
description = "A minimal HTTP server in Python, implemented using the standard http.server module"
authors = [
    "Nadezhda_Popova <nadezhdapopova13@yandex.ru>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
python-dotenv = ">=1.1.1,<2.0.0"
requests = ">=2.32.5,<3.0.0"
logging = ">=0.4.9.6,<0.5.0.0"


[build-system]
requires = "poetry-core>=2.0.0,<3.0.0"
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.17.1"
isort = "^6.0.1"

[tool.mypy]
disallow_untyped_defs=true
warn_return_any=true
exclude = """
/(
    .git
    .venv
    venv
)/
"""

[tool.black]
line-length = 119
exclude = """
/(
    .git
    .venv
    venv
)/
"""

[tool.isort]
line_length = 119
multi_line_output = 3